feat(ipc-shm): macOS 向け inline-tier IPC primitives を追加
`ring_consumer` を OS 別 backend (linux: memfd_create / macos: shm_open +
shm_unlink) に分割し、共通の SPSC consumer ロジックは `common` module に
括り出した。`mmap` 呼び出しは `shared::map_shared_fd` 1 箇所に集中させて
unsafe を最小化する。`fd_socket` は SCM_RIGHTS の意味論が両 OS で同一の
ため共通実装として両 OS でビルドする。

macOS では `/midori-<pid>-<nanos>-<counter>` 形式のユニーク名で `shm_open`
し、ftruncate / mmap 完了直後に `shm_unlink` で名前空間からエントリを除去
する POSIX 標準イディオムを採用。

公開 surface (`RingConsumer`, `CreateError`, `send_fd`, `recv_fd`) は当面
Linux のみで再エクスポートを維持する。macOS でも内部 module はビルド /
テストされ CI macos-latest job で挙動が検証されるが、Bridge runtime 層の
cfg ゲート整理は別 Issue で実施するため公開解禁はそこまで保留する。
