+++ crates/midori-runtime/src/lib.rs
@@ -22,11 +22,13 @@
 //! `mmap` and `SCM_RIGHTS` reception stay confined to that crate, and this
 //! crate can keep `unsafe_code = "forbid"` via `lints.workspace = true`.
 //!
-//! 現状 inline-tier IPC は Linux 専用（`memfd_create(2)` 依存、macOS は
-//! `shm_open(2)` ベース、Windows は `CreateFileMapping` ベースで実装される
-//! 予定だが未実装）。それまでは [`ring_ingest`] および `midori_ipc_shm` の
-//! Linux 専用 API を `cfg(target_os = "linux")` で囲って他 OS では参照しない
-//! 運用。
+//! 現状 inline-tier IPC の **公開 API** は Linux 専用（`memfd_create(2)`
+//! 依存）。macOS 向け実装 (`shm_open(2)` + `shm_unlink(2)` ベース) は
+//! `midori-ipc-shm` 内部に存在しビルド / テストされているが、Bridge
+//! runtime 層の cfg ゲート整理が完了するまで公開 surface としての解禁は
+//! 保留している。Windows backend (`CreateFileMapping` ベース) は未実装。
+//! それまでは [`ring_ingest`] および `midori_ipc_shm` の公開 API を
+//! `cfg(target_os = "linux")` で囲って他 OS では参照しない運用。
 //!
 //! The CLI dispatch layer remains binary-private until there is a concrete
 //! need to expose it.
